Refactor: FSD 폴더 구조 변경 (components, assets 제외)
