+++ src/app/index.jsx
@@ -5,15 +5,15 @@ import {
   useLocation,
   Navigate,
 } from 'react-router-dom';
-import Home from './pages/Home';
-import TrackInfo from './pages/TrackInfo';
+import Home from './pages/home';
+import TrackInfo from './pages/info';
 import NotFound from './pages/NotFound';
 import Search from './pages/Search';
-import Splash from './pages/Splash';
-import Header from './layouts/Header';
-import Footer from './layouts/Footer';
-import Login from './pages/Login';
-import MyPage from './pages/My';
+import Splash from './pages/splash';
+import Header from './widgets/navigation/Header';
+import Footer from './widgets/navigation/Footer';
+import Login from './pages/login';
+import MyPage from './pages/my';
 import UploadSection from './components/UploadSection';
 
 function App() {
