+++ src/entities/course/mock.js
