+++ src/pages/splash/index.jsx
