+++ src/shared/api/apiClient.js
@@ -1,0 +1,11 @@
+import axios from 'axios';
+import { BASE_URL } from '../../constants/api';
+
+const apiClient = axios.create({
+  baseURL: BASE_URL,
+  headers: {
+    'Content-Type': 'application/json',
+  },
+});
+
+export default apiClient;
