+++ src/widgets/navigation/Footer.jsx
